[Gold IV] Title: 행렬 제곱, Time: 0 ms, Memory: 2024 KB -BaekjoonHub

+++ 백준/Gold/10830. 행렬 제곱/행렬 제곱.cc
@@ -1,8 +1,8 @@
 #include <iostream>
 #include <vector>
 using namespace std;
-vector<vector<long>> multi(vector<vector<long>> &a, vector<vector<long>> &b, int n) {
-    vector<vector<long>> m(n, vector<long>(n, 0));
+vector<vector<int>> multi(vector<vector<int>> &a, vector<vector<int>> &b, int n) {
+    vector<vector<int>> m(n, vector<int>(n, 0));
     for (int i = 0; i < n; i++) {
         for (int j = 0; j < n; j++) {
             for (int k = 0; k < n; k++)
@@ -11,24 +11,24 @@ vector<vector<long>> multi(vector<vector<long>> &a, vector<vector<long>> &b, int
     }
     return m;
 }
-vector<vector<long>> square(vector<vector<long>> &x, long long exp, int n) {
+vector<vector<int>> square(vector<vector<int>> &x, long long exp, int n) {
     if (exp == 1)
         return x;
-    vector<vector<long>> temp = square(x, exp / 2, n);
-    vector<vector<long>> t = multi(temp, temp, n);
+    vector<vector<int>> temp = square(x, exp / 2, n);
+    vector<vector<int>> t = multi(temp, temp, n);
     return exp % 2 == 0 ? t : multi(t, x, n);
 }
 int main(void) {
     int n;
     long long b;
     cin >> n >> b;
-    vector<vector<long>> a(n, vector<long>(n, 0));
+    vector<vector<int>> a(n, vector<int>(n, 0));
     for (int i = 0; i < n; i++) {
         for (int j = 0; j < n; j++) {
             cin >> a[i][j];
         }
     }
-    vector<vector<long>> r = square(a, b, n);
+    vector<vector<int>> r = square(a, b, n);
     for (int i = 0; i < n; i++) {
         for (int j = 0; j < n; j++) {
             cout << r[i][j] % 1000 << ' ';
